2007-08-04  Thomas Koenig  <tkoenig@gcc.gnu.org>
	PR fortran/32770
	* gfortran.dg/streamio_8.f90:  Adjust so test case passes
	for -fdefault-integer-8 and -fdefault-real-8.
	* gfortran.dg/streamio_10.f90:  Likewise.
	* gfortran.dg/sizeof.f90:  Likewise.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@127212 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gfortran.dg/streamio_10.f90
@@ -4,7 +4,7 @@
 ! Test case derived from that given in PR by Steve Kargl.
 program stream_io_10
   implicit none
-  integer :: a(4), b(4)
+  integer(kind=4) :: a(4), b(4)
   integer(kind=8) :: thepos
   a = (/ 1, 2, 3, 4 /)
   b = a
